requirements.txt
streamlit==0.72.0
opencv_python==4.1.1.26
numpy==1.17.3
tensorflow_gpu==2.0.0
Pillow==8.0.1
tensorflow==2.4.0
